+++ app/Module/Users/Controller/UsersController.php
@@ -36,7 +36,8 @@ class UsersController extends UsersAppController
 	 * 
 	 * @return void
 	 */
-	public function admin_index(){
+	public function admin_index()
+	{
 		$this->paginate['User']['order'] = 'User.id Desc';
 		$this->set('users', $this->paginate('User'));
 	}
@@ -51,11 +52,11 @@ class UsersController extends UsersAppController
 		if ( !empty( $this->request->data ) ) {
 			$this->User->create();			
 			if ( $this->User->save( $this->request->data ) ) {
-				$this->Session->setFlash(__d('users', 'User saved.'), 'flash_success');
-				$this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__d('users', 'User saved.'), 'flash_success', array('plugin' => 'Pukis'));
+				$this->ajaxRedirect('/admin/users/users/index');
 			}
 		}
-		$roles = $this->Role->find('list');
+		$roles = $this->Role->find('list', array('conditions' => "Role.id >= '". $this->Auth->user('role_id') ."'"));
 		$this->set(compact('roles'));
 	}
 	
@@ -65,18 +66,34 @@ class UsersController extends UsersAppController
 	 * @param $id User ID
 	 * @return void
 	 */
-	public function admin_edit( $id = null ){
+	public function admin_edit( $id = null )
+	{
 		if ( !$id ) {
-			$this->Session->setFlash(__d('users', 'Invalid ID'), 'flash_error');
+			$this->Session->setFlash(__d('users', 'Invalid ID'), 'flash_error', array('plugin' => 'Pukis'));
 			$this->redirect(array('action' => 'index'));
 		}
-		if ( !empty( $this->request->data ) ) {
-			if ( $this->User->save($this->request->data) ) {
-				$this->Session->setFlash(__d('users', 'User was saved.'), 'flash_success');
+		
+		$editedUser = $this->User->find('first', array('conditions' => array('id' => $id), 'contain' => false));
+		
+		if ($this->Auth->user('id') != $id) {
+			if ($this->Auth->user('role_id') >= $editedUser['User']['role_id']) {
+				$this->Session->setFlash(__d('users', 'Cannot edit your own or upper class!'), 'flash_error', array('plugin' => 'Pukis'));
+				$this->ajaxRedirect('/admin/users/users/index');
 			}
 		}
-		$this->request->data = $this->User->read(null, $id);
-		$roles = $this->Role->find('list');
+		
+		if ( !empty( $this->request->data ) ) {
+			if ( $this->User->save($this->request->data) ) {
+				$this->Session->setFlash(__d('users', 'User was saved.'), 'flash_success', array('plugin' => 'Pukis'));
+				$this->ajaxRedirect('/admin/users/users/index');
+			}
+			
+			$this->Session->setFlash(__d('users', 'Unknown error occured!'), 'flash_error', array('plugin' => 'Pukis'));
+			$this->ajaxRedirect('/admin/users/users/index');
+		}
+		
+		$this->request->data = $editedUser;
+		$roles = $this->Role->find('list', array('conditions' => "Role.id >= '". $this->Auth->user('role_id') ."'"));
 		$this->set(compact('roles'));
 	}
 	
@@ -89,33 +106,27 @@ class UsersController extends UsersAppController
 	{
 		
 		if ($this->Auth->loggedIn()) {
-			//return $this->ajaxRedirect($this->Auth->redirect());
+			$this->Session->setFlash(__d('users', 'You are still have login'), 'flash_info', array('plugin' => 'Pukis'));
+			return $this->ajaxRedirect($this->Auth->redirect());
 		}
 		
 		if ( $this->request->is('post') ) {
-// 			if ( $this->Auth->login() ) {
-// 				return $this->ajaxRedirect($this->Auth->redirect());
-// 			}
-			
 			if ($this->Auth->login()) {
-			
 				// did they select the remember me checkbox?
 				if ($this->request->data['User']['remember_me'] == 1) {
-					// remove "remember me checkbox"
 					unset($this->request->data['User']['remember_me']);
-					
-					// hash the user's password
-					$this->request->data['User']['password'] = $this->Auth->password($this->request->data['User']['password']);
-			
-					// write the cookie
-					$this->Cookie->write('remember_me_cookie', $this->request->data['User'], true, '2 weeks');
+    	            $this->request->data['User']['password'] = $this->Auth->password($this->request->data['User']['password']);
+	                // write the cookie
+                	$this->Cookie->write('Pukis.token', $this->request->data['User'], true, Configure::read('AuthAccess.duration'));                	
 				}
-			
-				return $this->redirect($this->Auth->redirect());
 				
+				$this->Session->setFlash(__d('users', 'Login Success'), 'flash_success', array('plugin' => 'Pukis'));
+				return $this->ajaxRedirect($this->Auth->redirect());
 			}
+			
+			$this->Session->setFlash(__d('users', 'Login Error'), 'flash_error', array('plugin' => 'Pukis'));
+			return $this->ajaxRedirect($this->Auth->redirect());
 		}
-		
 	}
 	
 	/**
@@ -125,9 +136,13 @@ class UsersController extends UsersAppController
 	 */
 	public function admin_logout()
 	{
-		if ( $this->Auth->logout() ) {
+		if ($this->Auth->logout()) {
+			$this->Session->setFlash(__d('users', 'Logout Success'), 'flash_success', array('plugin' => 'Pukis'));
 			return $this->ajaxRedirect($this->Auth->redirect());
 		}
+		
+		$this->Session->setFlash(__d('users', 'Unknown error occured!'), 'flash_error', array('plugin' => 'Pukis'));
+		$this->ajaxRedirect('/admin/users/users/index');
 	}
 	
 	/**
@@ -136,17 +151,21 @@ class UsersController extends UsersAppController
 	 * @param  $id User ID
 	 * @return void
 	 */
-	public function admin_change_password( $id = null )
+	public function admin_change_password($id = null)
 	{
 		if ( !$id ) {
-			$this->Session->setFlash(__d('users', 'Invalid ID'), 'flash_error');
+			$this->Session->setFlash(__d('users', 'Invalid ID'), 'flash_error', array('plugin' => 'Pukis'));
 			$this->redirect(array('action' => 'index'));
 		}
+		
 		if ( !empty( $this->request->data ) ) {
 			if ( $this->User->save($this->request->data) ) {
-				$this->Session->setFlash(__d('users', 'password was saved.'), 'flash_success');
-				$this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__d('users', 'password was saved.'), 'flash_success', array('plugin' => 'Pukis'));
+				$this->ajaxRedirect('/admin/users/users/index');
 			}
+			
+			$this->Session->setFlash(__d('users', 'Unknown error occured!'), 'flash_error', array('plugin' => 'Pukis'));
+			$this->ajaxRedirect('/admin/users/users/index');
 		}
 	}
 	
@@ -156,23 +175,56 @@ class UsersController extends UsersAppController
 	 * @param $id User ID
 	 * @return void
 	 */
-	public function admin_delete( $id = null ){
-		if ( !$id ) {
-			$this->Session->setFlash(__d('users', 'Invalid ID!'), 'flash_error');
-			$this->redirect(array('action' => 'index'));
+	public function admin_delete($id = null, $confirm = 0)
+	{	
+		if (!$id) {
+			$this->Session->setFlash(__d('users', 'Invalid ID!'), 'flash_error', array('plugin' => 'Pukis'));
+			$this->ajaxRedirect('/admin/users/users/index');
 		}
-		if ( $this->User->delete( $id ) ) {
-			$this->Session->setFlash(__d('users', 'User was deleted.'), 'flash_success');
-			$this->redirect(array('action' => 'index'));
+		
+		if ($this->Auth->user('id') == $id) {
+			$this->Session->setFlash(__d('users', 'Cannot delete your own Id!'), 'flash_error', array('plugin' => 'Pukis'));
+			$this->ajaxRedirect('/admin/users/users/index');
 		}
+		
+		$deletedUser = $this->User->find('first',array('conditions' => array('id' => $id), 'contain' => false));
+		if ($this->Auth->user('role_id') >= $deletedUser['User']['role_id']) {
+			$this->Session->setFlash(__d('users', 'Cannot delete your own or upper class!'), 'flash_error', array('plugin' => 'Pukis'));
+			$this->ajaxRedirect('/admin/users/users/index');
+		}
+		
+		if($confirm == 1) {
+			if ( $this->User->delete($id) ) {
+				$this->Session->setFlash(__d('users', 'User was deleted.'), 'flash_success', array('plugin' => 'Pukis'));
+				$this->ajaxRedirect('/admin/users/users/index');	
+			}
+			
+			$this->Session->setFlash(__d('users', 'Unknown error occured!'), 'flash_error', array('plugin' => 'Pukis'));
+			$this->ajaxRedirect('/admin/users/users/index');
+		}
+		
+		$key = 'id';
+		$this->set(compact('key', 'id'));
 	}
 		
 	/**
 	 *  Install
 	 */
-	public function admin_install(){
+	public function admin_install()
+	{
 		$this->AclAco->create(array('parent_id' => null, 'alias' => 'controllers'));
 		$this->AclAco->save();
+	}
+	
+	/**
+	 * Generate login token to be saved in login table and cookies
+	 * this is used to match login token and prevent hack to cookies
+	 */
+	private function _generateLoginToken()
+	{
+		$userId = $this->Auth->user('id');
+		return md5($userId . date('Y m d H:i:s', strtotime('+2 Weeks')));
+		
 	}
 }
  ?>
